requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0
# ou PyMySQL==1.1.0 (alternativa)

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
argon2-cffi>=21.3.0

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Integration
openai==1.3.5
anthropic==0.7.0

# Email
python-dotenv==1.0.0
# Para produção: sendgrid ou mailgun

# Utilidades
python-dateutil==2.8.2
httpx==0.25.2

# CORS
fastapi-cors==0.0.6